feat(modules): modules own their admin APIs and wire types
@kroma/core knew what an indexer, a download client, a VPN and a naming
template were: 35 facade methods and ~40 zod schemas for routes that
module sidecars serve. All of it moves into the modules.

The client keeps one generic capability: `client.module(id)` binds a
{get, post, put, delete} surface to /api/admin/m/<id> — no module names,
JSON bodies typed correctly with the content type axum's Json extractor
insists on (covered by module-api.test, since nothing else would have
caught the 415s). JSON_HEADERS is hoisted into base.ts; ten files carried
their own copy.

The SDK grows the authoring surface:
- ModuleScope / useModuleApi: the host mounts each page with its id, so a
  module's own API never names the module.
- moduleApiHook(factory): a module's typed API is one expression — no
  React plumbing, no useMemo, no id. The (id, factory) overload exists
  for the one legitimate id: an API other packages call, published once
  next to its methods (acquisition, driven by the torrents manual grab).

Each module's wire types live in its schemas.ts, exported at
@kroma/module-<x>/schemas. Types the declared graph crosses are imported
by package name: torrents reads VpnStatusView from the vpn module it
optionally depends on. DownloadClientId moves with its only user;
brandedId is exported so a module can mint its own ids. Core keeps the
request/search flow (GrabBody, ScoredReleaseView, InteractiveSearchView)
and carries indexer ids as opaque brands.

The torrents -> acquisition frontend edge inverts the manifest graph
(acquisition dependsOn torrents); the entanglement predates this layout
and used to hide inside the monolithic client. It is documented at the
edge with the untangling path.
