+++ modules/tv.kroma.acquisition/ui/package.json
@@ -4,14 +4,17 @@
   "type": "module",
   "description": "KROMA Acquisition module (frontend half): the acquisition settings page. A settings-view module (no dedicated backend routes).",
   "exports": {
-    ".": "./src/index.ts"
+    ".": "./src/index.ts",
+    "./api": "./src/api.ts",
+    "./schemas": "./src/schemas.ts"
   },
   "scripts": {
     "typecheck": "tsc --noEmit",
     "build": "tsc --noEmit"
   },
   "dependencies": {
-    "@kroma/module-sdk": "workspace:*"
+    "@kroma/module-sdk": "workspace:*",
+    "zod": "^4.4.3"
   },
   "peerDependencies": {
     "react": ">=18"
