+++ modules/tv.kroma.indexer/ui/src/index.ts
@@ -4,7 +4,7 @@ import { lazy } from 'react';
 // The Indexers module (frontend half). id / version / dependsOn come from the
 // shared module.json (same file the backend reads). It contributes the full
 // Indexers admin page into the Acquisition sidebar group; the paired
-// IndexersModule ServerModule gates the /api/admin/indexers routes, so disabling
+// IndexersModule ServerModule gates its admin routes, so disabling
 // the module removes the page and its routes together.
 export const indexerModule = defineModule({
   pages: [
