+++ modules/tv.kroma.remote/ui/package.json
@@ -4,7 +4,8 @@
   "type": "module",
   "description": "KROMA Remote access module (frontend half): the Remote access admin page (share URL + managed Cloudflare Tunnel). Paired with the RemoteModule ServerModule via the shared module.json.",
   "exports": {
-    ".": "./src/index.ts"
+    ".": "./src/index.ts",
+    "./schemas": "./src/schemas.ts"
   },
   "scripts": {
     "typecheck": "tsc --noEmit",
