+++ modules/tv.kroma.torrents/ui/package.json
@@ -4,18 +4,22 @@
   "type": "module",
   "description": "KROMA Downloads module (frontend half): the Downloads admin page (live queue + download-clients). Paired with the kroma-torrent backend at ../server via the shared module.json.",
   "exports": {
-    ".": "./src/index.ts"
+    ".": "./src/index.ts",
+    "./schemas": "./src/schemas.ts"
   },
   "scripts": {
     "typecheck": "tsc --noEmit",
     "build": "tsc --noEmit"
   },
   "dependencies": {
+    "@kroma/module-acquisition": "workspace:*",
     "@kroma/module-sdk": "workspace:*",
+    "@kroma/module-vpn": "workspace:*",
     "@radix-ui/react-dropdown-menu": "^2.1.24",
     "@tabler/icons-react": "^3.44.0",
     "@tanstack/react-router": "^1.170.16",
-    "react-call": "^2.0.2"
+    "react-call": "^2.0.2",
+    "zod": "^4.4.3"
   },
   "peerDependencies": {
     "react": ">=18"
