+++ modules/tv.kroma.vpn/ui/package.json
@@ -4,7 +4,8 @@
   "type": "module",
   "description": "KROMA VPN module (frontend half): the VPN admin page (WireGuard bridge card + network settings). Paired with the VpnModule ServerModule via the shared module.json.",
   "exports": {
-    ".": "./src/index.ts"
+    ".": "./src/index.ts",
+    "./schemas": "./src/schemas.ts"
   },
   "scripts": {
     "typecheck": "tsc --noEmit",
@@ -12,7 +13,8 @@
   },
   "dependencies": {
     "@kroma/module-sdk": "workspace:*",
-    "@tabler/icons-react": "^3.44.0"
+    "@tabler/icons-react": "^3.44.0",
+    "zod": "^4.4.3"
   },
   "peerDependencies": {
     "react": ">=18"
