+++ packages/client/src/api.test.ts
@@ -349,11 +349,6 @@ describe('delegating methods issue the expected request', () => {
     ['updateLibrary', (c) => c.updateLibrary('x', {}), 'PATCH', '/admin/libraries/x'],
     ['deleteLibrary', (c) => c.deleteLibrary('x'), 'DELETE', '/admin/libraries/x'],
     ['scanLibrary', (c) => c.scanLibrary('x'), 'POST', '/admin/libraries/x/scan'],
-    ['adminNaming', (c) => c.adminNaming(), 'GET', '/admin/organize/naming'],
-    ['namingSample', (c) => c.namingSample({} as never), 'POST', '/admin/organize/sample'],
-    ['saveNaming', (c) => c.saveNaming({} as never), 'PUT', '/admin/organize/naming'],
-    ['organizePreview', (c) => c.organizePreview(), 'GET', '/admin/organize/preview'],
-    ['organizeApply', (c) => c.organizeApply(), 'POST', '/admin/organize/apply'],
   ];
 
   it.each(known)('%s -> %s %s', async (_name, call, method, path) => {
@@ -423,34 +418,6 @@ describe('delegating methods issue the expected request', () => {
     ['denyRequest', (c) => c.denyRequest('r1')],
     ['searchReleases', (c) => c.searchReleases('r1')],
     ['grabRelease', (c) => c.grabRelease('r1', {} as never)],
-    ['adminIndexers', (c) => c.adminIndexers()],
-    ['createIndexer', (c) => c.createIndexer({} as never)],
-    ['updateIndexer', (c) => c.updateIndexer('x', {} as never)],
-    ['deleteIndexer', (c) => c.deleteIndexer('x')],
-    ['testIndexer', (c) => c.testIndexer('x')],
-    ['adminIndexerDefinitions', (c) => c.adminIndexerDefinitions()],
-    ['indexerDefinitionDetail', (c) => c.indexerDefinitionDetail('x')],
-    ['syncIndexerDefinitions', (c) => c.syncIndexerDefinitions()],
-    ['adminDownloadClients', (c) => c.adminDownloadClients()],
-    ['createDownloadClient', (c) => c.createDownloadClient({} as never)],
-    ['updateDownloadClient', (c) => c.updateDownloadClient('x', {} as never)],
-    ['deleteDownloadClient', (c) => c.deleteDownloadClient('x')],
-    ['testDownloadClient', (c) => c.testDownloadClient('x')],
-    ['adminDownloads', (c) => c.adminDownloads()],
-    ['pauseDownload', (c) => c.pauseDownload('x')],
-    ['resumeDownload', (c) => c.resumeDownload('x')],
-    ['retryDownload', (c) => c.retryDownload('x')],
-    ['reannounceDownload', (c) => c.reannounceDownload('x')],
-    ['pauseAllDownloads', (c) => c.pauseAllDownloads()],
-    ['resumeAllDownloads', (c) => c.resumeAllDownloads()],
-    ['reannounceDownloads', (c) => c.reannounceDownloads()],
-    ['removeDownload', (c) => c.removeDownload('x')],
-    ['manualSearch', (c) => c.manualSearch('q')],
-    ['analyzeTorrent', (c) => c.analyzeTorrent('magnet:?x')],
-    ['manualAdd', (c) => c.manualAdd({} as never)],
-    ['adminVpn', (c) => c.adminVpn()],
-    ['saveVpn', (c) => c.saveVpn({} as never)],
-    ['testVpn', (c) => c.testVpn()],
     ['adminBrowseFolders', (c) => c.adminBrowseFolders()],
     ['adminServer', (c) => c.adminServer()],
     ['adminSessions', (c) => c.adminSessions()],
@@ -493,8 +460,6 @@ describe('delegating methods issue the expected request', () => {
     ['saveLlm', (c) => c.saveLlm({} as never)],
     ['llmModels', (c) => c.llmModels({} as never)],
     ['testLlm', (c) => c.testLlm({} as never)],
-    ['adminRemote', (c) => c.adminRemote()],
-    ['saveRemote', (c) => c.saveRemote({} as never)],
     ['logs', (c) => c.logs()],
     ['storyboard', (c) => c.storyboard('i1')],
   ];
