+++ packages/module-sdk/src/index.ts
@@ -4,26 +4,22 @@
 
 export * from '@kroma/admin-kit';
 export type {
-  ClientTestResult,
-  DownloadClientView,
-  DownloadView,
   EngineCapability,
-  IndexerDefinitionDetailView,
-  IndexerDefinitionView,
-  IndexerTestResult,
-  IndexerView,
-  ManualReleaseView,
+  KromaClient,
   MessageKey,
-  RemoteAccessView,
-  SaveDownloadClientBody,
-  SaveIndexerBody,
-  TorrentAnalysis,
-  TorrentFileView,
-  VpnTestResult,
+  ModuleApi,
+  ServerEvent,
 } from '@kroma/core';
 // The SSE client is renamed to avoid colliding with the SDK's own `KromaEvents`
 // event-map interface.
-export { apiErrorText, KromaEvents as KromaEventStream } from '@kroma/core';
+export {
+  apiErrorText,
+  brandedId,
+  IndexerId,
+  ItemId,
+  KromaEvents as KromaEventStream,
+  RequestId,
+} from '@kroma/core';
 export { useT } from '@kroma/ui';
 export { EmptyState } from '@kroma/ui/kit';
 export type { EventBus, EventKey } from './bus';
@@ -42,6 +38,7 @@ export type {
 } from './module';
 export type { ModuleNav, ModulePanel, ModuleRoute, ModuleStatus } from './registry';
 export { depEntries, ModuleRegistry } from './registry';
+export { ModuleScope, moduleApiHook, useModuleApi } from './scope';
 export type {
   Capability,
   CapabilityReq,
